+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improve http requests retry strategy. ([115](https://github.com/radiantearth/radiant-mlhub/pull/115))
+
 ### Fixed
 
 ### Deprecated